add forgot password, jpa specification based search

+++ src/main/java/com/cloud/CloudNest/services/impl/OtpServiceImpl.java
@@ -4,13 +4,19 @@ import com.cloud.CloudNest.dto.request.SendOtpRequest;
 import com.cloud.CloudNest.dto.request.VerifyOtpRequest;
 import com.cloud.CloudNest.dto.response.OtpResponse;
 import com.cloud.CloudNest.entities.EmailOtp;
+import com.cloud.CloudNest.entities.UserData;
+import com.cloud.CloudNest.enums.OtpType;
+import com.cloud.CloudNest.exception.UserAllReadyExistsException;
+import com.cloud.CloudNest.exception.UserInvalidInputException;
 import com.cloud.CloudNest.repository.EmailOtpRepository;
+import com.cloud.CloudNest.repository.UserDataRepository;
 import com.cloud.CloudNest.services.OtpService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.security.SecureRandom;
 import java.time.LocalDateTime;
@@ -21,52 +27,71 @@ public class OtpServiceImpl implements OtpService {
 
     private final EmailOtpRepository emailOtpRepository;
     private final JavaMailSender mailSender;
+    private final UserDataRepository userDataRepository;
 
     @Value("${spring.mail.username}")
     private String fromEmail;
 
+    @Transactional
     public OtpResponse sendOtp(SendOtpRequest request) {
 
         String email = request.getEmail();
 
-        // Delete previous OTP
-        emailOtpRepository.deleteByEmail(email);
+        UserData userData = userDataRepository.findByMail(email);
 
-        // Generate 6-digit OTP
-        String otp = String.valueOf(
-                100000 + new SecureRandom().nextInt(900000)
-        );
+        if (OtpType.REGISTRATION.equals(request.getOtpType()) && userData != null) {
+            throw new UserAllReadyExistsException("For mail id " + request.getEmail() + " User is present! Try with another mail Id");
+        }
 
-        // Create entity
-        EmailOtp emailOtp = EmailOtp.builder()
-                .email(email)
-                .otp(otp)
-                .expiryTime(LocalDateTime.now().plusMinutes(5))
-                .verified(false)
-                .createdAt(LocalDateTime.now())
-                .build();
+        // Forgot Password
+        if (request.getOtpType().equals(OtpType.FORGOT_PASSWORD) && userData == null) {
+            throw new UserInvalidInputException("Email is not registered");
+        }
 
-        // Save OTP in DB
-        emailOtpRepository.save(emailOtp);
+        try {
+            // Delete previous OTP
+            emailOtpRepository.deleteByEmail(email);
 
-        // Send email
-        SimpleMailMessage message = new SimpleMailMessage();
+            // Generate 6-digit OTP
+            String otp = String.valueOf(
+                    100000 + new SecureRandom().nextInt(900000)
+            );
 
-        message.setFrom(fromEmail);
-        message.setTo(email);
-        message.setSubject("CloudNest - Email Verification");
+            // Create entity
+            EmailOtp emailOtp = EmailOtp.builder()
+                    .email(email)
+                    .otp(otp)
+                    .expiryTime(LocalDateTime.now().plusMinutes(5))
+                    .verified(false)
+                    .createdAt(LocalDateTime.now())
+                    .build();
 
-        message.setText(
-                "Your OTP is: " + otp +
-                        "\n\nThis OTP will expire in 5 minutes."
-        );
+            // Save OTP in DB
+            emailOtpRepository.save(emailOtp);
 
-        mailSender.send(message);
+            // Send email
+            SimpleMailMessage message = new SimpleMailMessage();
+
+            message.setFrom(fromEmail);
+            message.setTo(email);
+            message.setSubject("CloudNest - Email Verification");
+
+            message.setText(
+                    "Your OTP is: " + otp +
+                            "\n\nThis OTP will expire in 5 minutes."
+            );
+
+            mailSender.send(message);
+        } catch (Exception e) {
+            e.printStackTrace();
+            throw new RuntimeException("There some Issue. Try later");
+        }
 
         return new OtpResponse("OTP sent successfully");
     }
 
 
+    @Transactional
     public OtpResponse verifyOtp(VerifyOtpRequest request) {
 
         EmailOtp emailOtp = emailOtpRepository
